Move INotifyRecursive to its own file
